Move the executor to services and test fixture functions to conftest
