+++ h/views/api/bulk.py
@@ -1,15 +1,10 @@
 from itertools import chain
-from logging import getLogger
 
 from pyramid.response import Response
 
 from h.h_api.bulk_api import BulkAPI
-from h.h_api.bulk_api.executor import AutomaticReportExecutor
-from h.h_api.enums import DataType
-from h.h_api.exceptions import InvalidDeclarationError
+from h.services.bulk_executor._executor import  AuthorityCheckingExecutor
 from h.views.api.config import api_config
-
-LOG = getLogger(__name__)
 
 
 class FakeBulkAPI(BulkAPI):
@@ -30,41 +25,6 @@ class FakeBulkAPI(BulkAPI):
             yield b'["Line 2"]\n'
         else:
             return None
-
-
-class AuthorityCheckingExecutor(AutomaticReportExecutor):
-    """A bulk executor which checks the authority."""
-
-    def __init__(self, authority="lms.hypothes.is"):
-        self.effective_user = None
-        self.authority = authority
-
-    def configure(self, config):
-        self._assert_authority("effective user", config.effective_user, embedded=True)
-
-        self.effective_user = config.effective_user
-
-    def execute_batch(self, command_type, data_type, default_config, batch):
-        for command in batch:
-            self._check_authority(data_type, command.body)
-
-        return super().execute_batch(command_type, data_type, default_config, batch)
-
-    def _assert_authority(self, field, value, embedded=False):
-        if embedded and value.endswith(f"@{self.authority}"):
-            return
-
-        if value == self.authority:
-            return
-
-        raise InvalidDeclarationError(
-            f"The {field} '{value}' does not match the expected authority"
-        )
-
-    def _check_authority(self, data_type, body):
-        if data_type in (DataType.USER, DataType.GROUP):
-            self._assert_authority("authority", body.attributes["authority"])
-            self._assert_authority("query authority", body.query["authority"])
 
 
 @api_config(
